+++ src/main/java/com/cpz/sim/datacenter/workload/WorkloadSource.java
@@ -9,7 +9,7 @@ import com.cpz.sim.foundation.time.SimulationTick;
 @FunctionalInterface
 public interface WorkloadSource {
 
-    float getUtilization(Server server, SimulationTick tick);
+    double getUtilization(Server server, SimulationTick tick);
 
     default void reset() {
     }
