+++ src/test/java/com/cpz/sim/datacenter/integration/TemperatureSimulationIntegrationTest.java
@@ -291,7 +291,7 @@ class TemperatureSimulationIntegrationTest {
         );
     }
 
-    private static WorkloadSource fixedWorkload(float utilization) {
+    private static WorkloadSource fixedWorkload(double utilization) {
         return (server, tick) -> utilization;
     }
